+++ docs/changes.rst
@@ -5,7 +5,7 @@ Weblate 5.17
 
 .. rubric:: New features
 
-* Added :setting:`WEBSITE_ALERTS_ENABLED` setting to allow disabling website availability checks and alerts.
+* Added :setting:`WEBSITE_ALERTS_ENABLED` setting to allow disabling project website availability checks and alerts.
 * :doc:`/formats/asciidoc` is now a supported file format.
 
 .. rubric:: Improvements